(PC-32696)[PRO] refactor: mv OfferLocation component in DetailsAndInformations dir since its only used there
