+++ pro/src/components/IndividualOffer/UsefulInformationScreen/UsefulInformationForm/UsefulInformationForm.tsx
@@ -13,7 +13,7 @@ import { useCurrentUser } from 'commons/hooks/useCurrentUser'
 import { OfferRefundWarning } from 'components/Banner/OfferRefundWarning'
 import { WithdrawalReminder } from 'components/Banner/WithdrawalReminder'
 import { FormLayout } from 'components/FormLayout/FormLayout'
-import { OfferLocation } from 'pages/IndividualOffer/components/OfferLocation/OfferLocation'
+import { OfferLocation } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/components/OfferLocation/OfferLocation'
 import { Checkbox } from 'ui-kit/form/Checkbox/Checkbox'
 import { CheckboxGroup } from 'ui-kit/form/CheckboxGroup/CheckboxGroup'
 import { RadioGroup } from 'ui-kit/form/RadioGroup/RadioGroup'
