+++ pro/src/components/IndividualOffer/UsefulInformationScreen/__specs__/utils.spec.ts
@@ -10,7 +10,7 @@ import {
   subcategoryFactory,
   venueListItemFactory,
 } from 'commons/utils/individualApiFactories'
-import { OFFER_LOCATION } from 'pages/IndividualOffer/components/OfferLocation/constants'
+import { OFFER_LOCATION } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/components/OfferLocation/constants'
 
 import { DEFAULT_USEFULL_INFORMATION_INTITIAL_VALUES } from '../constants'
 import {
