+++ pro/src/components/IndividualOffer/UsefulInformationScreen/utils.ts
@@ -11,7 +11,7 @@ import {
 } from 'commons/core/Offers/constants'
 import { isOfferSynchronized } from 'commons/core/Offers/utils/typology'
 import { AccessibilityEnum } from 'commons/core/shared/types'
-import { OFFER_LOCATION } from 'pages/IndividualOffer/components/OfferLocation/constants'
+import { OFFER_LOCATION } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/components/OfferLocation/constants'
 import { FORM_DEFAULT_VALUES } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferInformationsLegacy/InformationsScreen/IndividualOfferForm/commons/constants'
 import { computeAddressDisplayName } from 'repository/venuesService'
