+++ pro/src/components/IndividualOffer/UsefulInformationScreen/validationSchema.ts
@@ -1,7 +1,7 @@
 import * as yup from 'yup'
 
 import { WithdrawalTypeEnum } from 'apiClient/v1'
-import { validationSchema as offerLocationSchema } from 'pages/IndividualOffer/components/OfferLocation/validationSchema'
+import { validationSchema as offerLocationSchema } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/components/OfferLocation/validationSchema'
 
 // FIX ME: this regex is subject to backtracking which can lead to "catastrophic backtracking", high memory usage and slow performance
 // we cannot use the yup url validation because we need to allow {} in the url to interpolate some data
