+++ pro/src/pages/IndividualOffer/IndividualOfferDetailsAndInformations/components/OfferLocation/constants.ts
