+++ pro/src/pages/IndividualOffer/commons/serializers/serializePatchOffer.ts
@@ -9,7 +9,7 @@ import { isAllocineOffer } from 'commons/core/Providers/utils/localProvider'
 import { AccessibilityEnum } from 'commons/core/shared/types'
 import { trimStringsInObject } from 'commons/utils/trimStringsInObject'
 import { IndividualOfferFormValues } from 'pages/IndividualOffer/commons/types'
-import { OFFER_LOCATION } from 'pages/IndividualOffer/components/OfferLocation/constants'
+import { OFFER_LOCATION } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/components/OfferLocation/constants'
 
 export const serializeExtraData = (
   formValues: Partial<IndividualOfferFormValues>
